fix: resolve clippy warnings for CI
Apply manual_find and unnecessary_cast fixes, and move test modules to
the end of proxy_keys.rs and assistant_mcp.rs per items_after_test_module.

+++ src/api/control/mod.rs
@@ -19107,7 +19107,7 @@ Investigate <service/> failures.
                         tokio::time::sleep(std::time::Duration::from_millis(50)).await;
                     }
                     Err(broadcast::error::RecvError::Lagged(n)) => {
-                        lagged = lagged.saturating_add(n as u64);
+                        lagged = lagged.saturating_add(n);
                     }
                     Err(broadcast::error::RecvError::Closed) => break,
                 }
